Fix test dependencies
