+++ datacapturing/build.gradle
@@ -110,6 +110,7 @@ dependencies {
     androidTestImplementation "androidx.test:rules:$rootProject.ext.rulesVersion"
 
     // Dependencies for local unit tests
+    testImplementation project(":testutils")
     // - If Junit symbols are not resolvable in IntelliJ, make sure Build Variant is set to debug
     // - Loading another dependency (e.g. module) only it's production dependencies (compile) are loaded but not other dependencies (e.g. testCompile)
     testImplementation "junit:junit:$rootProject.ext.junitVersion"
