+++ persistence/build.gradle
@@ -37,6 +37,7 @@ dependencies {
     implementation project(":utils")
 
     // Dependencies for local unit tests
+    testImplementation project(":testutils")
     // - If Junit symbols are not resolvable in IntelliJ, make sure Build Variant is set to debug
     // - Loading another dependency (e.g. module) only it's production dependencies (compile) are loaded but not other dependencies (e.g. testCompile)
     testImplementation "junit:junit:$rootProject.ext.junitVersion"
